se crea archivo constants dentro del componente TechStack para separar el arreglo de la pagina principal y se crea archivo SingleTechStack donde se renderizan las imagenes con su nombre
