+++ components/TechStack/SingleTechStack.tsx
@@ -1,0 +1,26 @@
+import Image from "next/image";
+import { TechStackType } from "@/types/techstack";
+
+
+const SingleTechStack = ({ techstack }: { techstack: TechStackType }) => {
+    const { image, name } = techstack;
+  
+    return (
+  
+      <div className="mx-3 flex w-full max-w-[160px] items-center justify-center py-[15px] sm:mx-4 lg:max-w-[130px] xl:mx-6 xl:max-w-[150px] 2xl:mx-8 2xl:max-w-[160px]">
+        <a
+          href="/"
+          target="_blank"
+          rel="nofollow noreferrer"
+          className="relative h-10 w-full opacity-70 grayscale transition hover:opacity-100 hover:grayscale-0 dark:opacity-60 dark:hover:opacity-100"
+        >
+          <Image src={image} alt={name} fill />
+        </a>
+  
+  
+      </div>
+    
+    );
+  };
+
+export default SingleTechStack;
