+++ components/TechStack/constants.tsx
@@ -1,0 +1,82 @@
+import { TechStackType } from "@/types/techstack";
+
+export const TechStackData: TechStackType[] = [
+    {
+      id: 1,
+      name: "css",
+      image: "/images/techstack/css-3.svg",
+    },
+    {
+      id: 2,
+      name: "dart",
+      image: "/images/techstack/dart.svg",
+    },
+    {
+      id: 3,
+      name: "gopher",
+      image: "/images/techstack/go-8.svg",
+    },
+    {
+      id: 4,
+      name: "html",
+      image: "/images/techstack/html-1.svg",
+    },
+    {
+      id: 5,
+      name: "Java",
+      image: "/images/techstack/java-14.svg",
+    },
+    {
+      id: 6,
+      name: "javascript", 
+      image: "/images/techstack/logo-javascript.svg",
+    },
+    {
+      id: 7,
+      name: "cmas",
+      image: "/images/techstack/cmas.svg",
+    },
+    {
+      id: 8,
+      name: "csharp",
+      image: "/images/techstack/csharp.svg",
+    },
+    {
+      id: 9,
+      name: "php",
+      image: "/images/techstack/icons8-logo-php.svg",
+    },
+    {
+      id: 10,
+      name: "php",
+      image: "/images/techstack/python-5.svg",
+    },
+    {
+      id: 11,
+      name: "react",
+      image: "/images/techstack/react-2.svg",
+    },
+    {
+      id: 12,
+      name: "ruby",
+      image: "/images/techstack/ruby.svg",
+    },
+    {
+      id: 13,
+      name: "rust",
+      image: "/images/techstack/rust.svg",
+    },
+  
+    {
+      id: 14,
+      name: "typescript",
+      image: "/images/techstack/typescript-2.svg",
+    },
+    {
+      id: 15,
+      name: "swift",
+      image: "/images/techstack/swift-15.svg",
+    }
+  ];
+
+  export default TechStackData;
